Googletest export
Breaks Windows builds

PiperOrigin-RevId: 210434120
